refactor: rename store folder to stores
