+++ inertia/stores/collection_store.ts
