+++ inertia/stores/global_hotkeys_store.ts
